paralel with mpi
