+++ .gitignore
@@ -45,6 +45,7 @@
 # Kernel Module Compile Results
 *.mod*
 *.cmd
+saida.txt
 .tmp_versions/
 modules.order
 Module.symvers
