--- src/prime.h
@@ -1,1 +1,0 @@
-int div_num(int x);
